Added carousel and images
Using the react-carousel component, I got one working but the issue is it clips outside the page and looks odd overall. I'll switch to Embla Carousel in a later commit.

Images have also been added to public folder along with json files that link to all the images and contain metadata for the carousels.